+++ changes.md
@@ -5,6 +5,7 @@ DartMob Changes
 
 * Implemented DeckPanel. Added example.
 * Implemented HtmlTable.
+* Implemented FlexTable. Added example.
 
 **January 2, 2013